Simplify headers, with a common class, and ensure jupyterhub version header is present
